changed the trace partitionning abstraction
given a conditional [c] and a program point [pt] at which [c] occurs,
traces are split between in three sets:
- [n_true] and [n_false] are the set of traces that went at least of
  through [c] at [pt], and such that [c] evaluated to, respectively,
  true and false the last time the trace went through it.
- [n_unkwn] are the set of traces that never went through [c] at [pt]

Note that we do this splitting according to (well-nested)
partitioning blocks. Actually, the condition:
"went through [c] at [pt]"
is
"went through [c] at [pt] in the current partitioning block"
For example, a partitioning block is open before stabilizing a while
statement, and is closed after stabilizing it.